feat(house): chips + optional input; API routes; histogram axes; collapsible log (mobile-friendly) feat(treasure): chosen-card highlight, disable roll during choice, end-game action; player chips fix: rename sim->house paths and templates; update config to pydantic-settings chore: update requirements and .gitignore
