+++ app/core/templates.py
@@ -1,0 +1,4 @@
+from fastapi.templating import Jinja2Templates
+from .config import settings
+
+templates = Jinja2Templates(directory=settings.TEMPLATE_DIR)
