+++ app/templates/base.html
@@ -1,0 +1,28 @@
+<!doctype html>
+<html>
+<head>
+  <meta charset="utf-8" />
+  <title>EDH Tools • Home</title>
+  <meta name="viewport" content="width=device-width, initial-scale=1" />
+  <link rel="stylesheet" href="/static/css/base.css">
+</head>
+<body>
+  {% include "_nav.html" %}
+
+  <div class="container">
+    <h1>EDH Tools</h1>
+
+    <div class="grid grid-2">
+      <a class="card card--clickable" href="/house" aria-label="Gamble: Puzzlebox Loop Helper">
+        <div style="font-weight:800; margin-bottom:4px;">Gamble</div>
+        <div class="muted">Puzzlebox Loop Helper</div>
+      </a>
+
+      <a class="card card--clickable" href="/treasure" aria-label="Treasure Cruise: Deck Importer">
+        <div style="font-weight:800; margin-bottom:4px;">Treasure Cruise</div>
+        <div class="muted">Deck Importer</div>
+      </a>
+    </div>
+  </div>
+</body>
+</html>
