+++ app/templates/treasure/session.html
@@ -1,0 +1,60 @@
+<!doctype html>
+<html>
+<head>
+  <meta charset="utf-8" />
+  <title>Treasure Cruise • Session</title>
+  <link rel="stylesheet" href="/static/css/base.css">
+</head>
+<body>
+  {% include "_nav.html" %}
+
+  <div class="container stack-12" id="tc-root" data-sid="{{ sid }}">
+    <div class="row" style="justify-content:space-between; align-items:center;">
+      <div class="muted" style="font-size:14px;">
+        Session code:
+        <code id="tc-sid">{{ sid }}</code>
+        <button class="btn btn--small" id="tc-sid-copy">Copy</button>
+      </div>
+      <div class="row" id="tc-players"></div>
+    </div>
+
+    <h1>Treasure Cruise • Session</h1>
+    <div class="muted" id="tc-turnlabel">Turn —</div>
+
+    <div class="card stack-12">
+      <div class="row" style="justify-content:space-between; align-items:center;">
+        <h2 style="margin:0;">Top of Deck</h2>
+        <div class="muted" style="font-size:12px;">latest reveal</div>
+      </div>
+      <div id="tc-deck" class="tc-cards"></div>
+    </div>
+
+    <div class="card stack-12">
+      <div class="row" style="justify-content:space-between; align-items:center;">
+        <h2 style="margin:0;">Active Player’s Cards</h2>
+        <div class="muted" style="font-size:12px;">auto-updates on turn</div>
+      </div>
+      <div id="tc-hand" class="tc-cards"></div>
+    </div>
+
+    <div class="card">
+      <div class="row" style="justify-content:space-between; align-items:center;">
+        <h2 style="margin:0;">Log</h2>
+        <button class="log__toggle" id="tc-log-toggle" type="button">Show all</button>
+      </div>
+      <div class="log__box" id="tc-logbox"></div>
+    </div>
+  </div>
+
+  <div class="container sticky-bottom">
+    <div class="actionbar">
+      <button class="btn" id="tc-roll">Pay — • Roll d6</button>
+      <button class="btn btn--ghost" id="tc-pass">Pass turn</button>
+      <button class="btn btn--danger btn--ghost" id="tc-end">End game</button>
+    </div>
+  </div>
+
+  <script src="/static/js/base.js"></script>
+  <script src="/static/js/treasure.js"></script>
+</body>
+</html>
