+++ app/web/home.py
@@ -1,0 +1,9 @@
+from fastapi import APIRouter, Request
+from fastapi.responses import HTMLResponse
+from app.core.templates import templates
+
+router = APIRouter()
+
+@router.get("/", name="home", response_class=HTMLResponse)
+def home(request: Request):
+    return templates.TemplateResponse("base.html", {"request": request})
