+++ requirements.txt
@@ -5,3 +5,5 @@ python-multipart==0.0.9
 pydantic==2.9.2
 requests==2.32.3
 psycopg[binary,pool]>=3.1
+pydantic>=2.0
+pydantic-settings>=2.2
