--- static/house.css
@@ -1,13 +1,0 @@
-/* House (Puzzlebox) specific tweaks */
-
-/* Keep the checkbox + label perfectly aligned */
-.checklist { display:flex; align-items:center; gap:10px; }
-.checklist input[type="checkbox"] { width: 18px; height: 18px; }
-.checklist label { margin: 0; }
-
-/* Slightly larger “k/v” pairs for readability */
-.k { font-size: 14px; color: var(--fg-dim); letter-spacing: .2px; }
-.v { font-size: 18px; font-weight: 800; margin: 2px 0 8px; }
-
-/* Keep sticky bar from looking empty on first load */
-#sticky .row { min-height: 28px; }
